segmentation/mask_rcnn/requirements.txt
lxml
matplotlib
numpy
tqdm
pycocotools
Pillow

